refactor: Reuse `@deprecated` import
